+++ blog/src/main/java/com/example/demo/model/BlogExample.java
@@ -115,42 +115,42 @@ public class BlogExample {
             return (Criteria) this;
         }
 
-        public Criteria andBlogIdEqualTo(Integer value) {
+        public Criteria andBlogIdEqualTo(Long value) {
             addCriterion("blog_id =", value, "blogId");
             return (Criteria) this;
         }
 
-        public Criteria andBlogIdNotEqualTo(Integer value) {
+        public Criteria andBlogIdNotEqualTo(Long value) {
             addCriterion("blog_id <>", value, "blogId");
             return (Criteria) this;
         }
 
-        public Criteria andBlogIdGreaterThan(Integer value) {
+        public Criteria andBlogIdGreaterThan(Long value) {
             addCriterion("blog_id >", value, "blogId");
             return (Criteria) this;
         }
 
-        public Criteria andBlogIdGreaterThanOrEqualTo(Integer value) {
+        public Criteria andBlogIdGreaterThanOrEqualTo(Long value) {
             addCriterion("blog_id >=", value, "blogId");
             return (Criteria) this;
         }
 
-        public Criteria andBlogIdLessThan(Integer value) {
+        public Criteria andBlogIdLessThan(Long value) {
             addCriterion("blog_id <", value, "blogId");
             return (Criteria) this;
         }
 
-        public Criteria andBlogIdLessThanOrEqualTo(Integer value) {
+        public Criteria andBlogIdLessThanOrEqualTo(Long value) {
             addCriterion("blog_id <=", value, "blogId");
             return (Criteria) this;
         }
 
-        public Criteria andBlogIdIn(List<Integer> values) {
+        public Criteria andBlogIdIn(List<Long> values) {
             addCriterion("blog_id in", values, "blogId");
             return (Criteria) this;
         }
 
-        public Criteria andBlogIdNotIn(List<Integer> values) {
+        public Criteria andBlogIdNotIn(List<Long> values) {
             addCriterion("blog_id not in", values, "blogId");
             return (Criteria) this;
         }
